BIOMOD_Modeling wrapper

BIOMOD_Modeling wrapper

biomodModelingWrapper(
  sp,
  responseVar,
  responseVarData,
  predictorVars,
  predictorVarsData,
  dir.name,
  BIOMOD_ModelingArgs = list(models = c("GLM", "MARS"), bm.options = NULL, CV.k = 5,
    CV.perc = 100, metric.eval = c("TSS", "ROC"), modeling.id = "test"),
  ...
)

Arguments

sp

character of species name to subset responseVarData table

responseVar

character vector, or list of character vectors, of response variable (column) to subset responseVarData table. If a list, it should be named according to sp, for subsetting.

responseVarData

a data.table or list of data.tables with environmental data. If a list, it should be named according to sp, for subsetting.#'

predictorVars

character vector, or list of character vectors, of environmental variables (columns) to subset predictorVarsData table. If a list, it should be named according to sp, for subsetting.

predictorVarsData

environmental data.

dir.name

passed to biomod2::BIOMOD_FormatingData

BIOMOD_ModelingArgs

a named list of arguments passed to biomod2::BIOMOD_Modeling

...

further arguments passed to reproducible::Cache