Activate serotiny after a (fire) disturbance

Activate serotiny after a (fire) disturbance

doSerotiny(
  burnedPixelCohortData,
  postFirePixelCohortData,
  postFireRegenSummary = NULL,
  species,
  sufficientLight,
  speciesEcoregion,
  currentTime,
  treedFirePixelTableSinceLastDisp,
  calibrate = FALSE
)

Arguments

burnedPixelCohortData

An expanded cohortData data.table with pixel-level cohort information on burnt pixels and the following (optional) columns: severity - fire severity in that pixel calculated based on fire behaviour properties; firetolerance - species-level fire tolerance; severityToleranceDif - the difference between severity and firetolerance.

postFirePixelCohortData

an empty cohortData-like table with columns: age, B, mortality, aNPPAct, and sumB removed; and pixelIndex added.

postFireRegenSummary

a data.table summarizing for which species serotiny/resprouting were activated and in how many pixels, for each year. Only necessary if calibrate = TRUE.

species

a data.table with species traits such as longevity, shade tolerance, etc. Must have column speciesCode, with species names/IDs. The following is a list of default trait columns:

"species" same as "speciesCode" -- species ID name

"speciesCode"

"Area" -- inherited from LANDIS-II default table, the Canadian ecoregion from which traits where derived. Not used during the simulation

"firetolerance" -- modulerelative (to other species) fire tolerance

"growthcurve" and "mortalityshape" -- growth curve shape parameters.

"longevity" -- maximum species age

"postfireregen" -- post-fire regeneration strategy ("serotiny", "resprout" or "none")

"resproutprob" -- probability of resprouting

"resproutage_min" -- minimum age at which species is capable of resprouting

"resproutage_max" -- maximum age at which species is capable of resprouting

"seeddistance_eff" -- effective dispersal distance

"seeddistance_max" -- maximum dispersal distance

"shadetolerance" -- relative (to other species) shade tolerance

"sexualmature" -- age at sexual maturity Known optional parameters added/needed by some modules (the user may add others for their own modules):

"inflationFactor" -- Biomass_speciesParameters module: inflation factor for maxB

"growthCurveSource" -- Biomass_speciesParameters module: how "growthcurve" was estimated

"mANPPproportion" -- Biomass_speciesParameters module: multiplication factor to calculate maxANPP from maxB

"thermokarsttol" -- Biomass_disturbances module: proportion of biomass surviving after thermokarst (i.e. permafrost thaw). Applied equally across cohorts. Parameters inherited from LANDIS-II default table, but not used in LandR at the moment:

"leaflongevity"

"wooddecayrate"

"leafLignin"

"hardsoft" Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015) for further detail.

sufficientLight

a data.frame (not data.table!) containing probability of establishment, given a site's light conditions (X0-X5) for each level of a species shade tolerance (1-5).

speciesEcoregion

A data.table with species-ecoregion-specific species trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical characteristics. The table should have at least the following columns: speciesCode and ecoregionGroup, character representation of species and ecoregion groups respectively, maxB the maximum biomass for the species in a given 'ecoregion', maxANPP the maximum aboveground net primary productivity and SEP the species establishment probability. May contain columns inflationFactor (used to adjust maxB) and mANPPproportion (used to calculate maxANPP).

currentTime

integer. The current simulation time obtained with time(sim)

treedFirePixelTableSinceLastDisp

a vector of pixels that burnt and were forested in the previous time step.

calibrate

logical. Determines whether to output postFirePixelCohortData. Defaults to FALSE

Value

A list of objects: postFirePixelCohortData, a data.table with the cohorts that undergo serotiny; serotinyPixel, a vector of pixels where serotiny was activated; postFireRegenSummary, the updated postFireRegenSummary, if calibrate = TRUE.