Create dummy inputs for test simulations

ecoregionMapis a raster of all the unique groupings.

makeDummyEcoregionMap(rasterToMatch)

makeDummyRawBiomassMap(rasterToMatch)

makeDummyStandAgeMap(rawBiomassMap)

makeDummyRstLCC(rasterToMatch)

makeDummyEcoregionFiles(ecoregionMap, rstLCC, rasterToMatch)

Arguments

rasterToMatch

A RasterLayer objects to use as the template for all subsequent raster operations (i.e., the one used throughout the simulation).

rawBiomassMap

a rawBiomassMap (e.g. the one used throughout the simulation)

ecoregionMap

The ecoregionMap, a raster of all the unique ecoregion groupings, which group together pixels of similar biophysical characteristics.

rstLCC

raster of land-cover class, e.g., LCC2005

Value

a RasterLayer object or, in the case of makeDummyEcoregionFiles, a list.

Details

rawBiomassMap is a raster of "raw" total stand biomass per pixel, with values between 100 and 20000 g/m^2.

standAgeMap is a raster of stand age per pixel (where biomass exists) with values between 1 and 300 years.

rstLCC is a raster land-cover class per pixel, with values between 1 and 5 that have no correspondence to any real land-cover classes.

ecoregionFiles uses dummy versions of ecoregionMap and rstLCC to create a list with two objects: the ecoregionMap and a table summarizing its information per pixelID. See ecoregionProducer (it uses ecoregionProducer internally).