Plot estimated maximum biomass by age

Plots a maximum biomass estimated at maximum 'cover' (or dominance) levels as a function of age.

ggplotMLL_maxB(
  mll,
  data,
  maxCover = 1L,
  xCovar = "age",
  plotTitle = NULL,
  nonLinModelQuoted,
  linModelQuoted,
  averageCovariates = TRUE,
  observedAge = FALSE,
  plotCIs = TRUE
)

Arguments

mll

a named list with outputs of an bbmle::mle2 call (the fitted non-linear model), from which coefficient values will be extracted. If several model outputs are provided all fitted models will be plotted, with plot labels corresponding to list names.

data

data for estimation of maximum biomass. Should contain at least an 'age' column. Note that other covariates will be averaged and 'cover' values will be replaced with the maximum cover value (maxCover). If mll is a list data is assumed to be the same for the two models.

maxCover

numeric. Value indicating maximum cover/dominance.

xCovar

the variable shown in the x axis. Defaults to age.

plotTitle

character. Passed to title in ggplot2::labs().

nonLinModelQuoted

a named list of non-linear equations as a call (quoted expression) passed to mle2(minuslog1). See bbmle::mle2(). Accepts equations with three parameters 'A', 'p' and 'k'. List names and length must the same as in mll.

linModelQuoted

A named list of lists of linear equations/modes relating each parameter ('A', 'p' and 'k') with a set of covariates. A call (quoted expression) passed to mle2(..., parameters). Note that for the purpose of tree growth, the linear equation determining 'A' should include a 'cover' predictor indicating the tree cover or dominance in the stand. Should be scaled between 0 and maxCover. List names and length must the same as in mll.

averageCovariates

should covariates other than age/cover be averaged for biomass predictions? If not, for each age (at maximum cover) there will be as many predictions as other covariate values. If observedAge == TRUE and averageCovariates == FALSE then the original data is used, with cover changed to maxCover.

observedAge

should observed age values be used, or should these be generated as round(seq(min(age), max(age)*1.5, length.out = 100), 0)? If observedAge == TRUE and averageCovariates == FALSE then the original data is used, with cover changed to maxCover.

plotCIs

should confidence intervals be calculated and plotted?