Load kNN species layers from online data repository

TODO: description needed

loadkNNSpeciesLayers(dPath, rasterToMatch, studyArea, sppEquiv,
  knnNamesCol = "KNN", sppEquivCol, thresh = 1, url, ...)

Arguments

dPath
path to the data directory

rasterToMatch
passed to prepInputs

studyArea
passed to prepInputs

sppEquiv
table with species name equivalencies between the kNN format and the final naming format. See data("sppEquivalencies_CA", "pemisc").

knnNamesCol
character string indicating the column in sppEquiv containing kNN species names. Default "KNN" for when sppEquivalencies_CA is used.

sppEquivCol
character string indicating the column in sppEquiv to use for final species names.

thresh
the minimum number of pixels where the species must have biomass > 0 to be considered present in the study area. Defaults to 1.

url
the source url for the data, passed to prepInputs

...
Additonal arguments passed to Cache and equivalentName.

Value

A raster stack of percent cover layers by species.