Create speciesEcoregion

Use statistically estimated maxB, maxANPP and establishment probabilities to generate speciesEcoregion table.

makeSpeciesEcoregion(
  cohortDataBiomass,
  cohortDataShort,
  cohortDataShortNoCover,
  species,
  modelCover,
  modelBiomass,
  successionTimestep,
  currentYear
)

Arguments

cohortDataBiomass

a subset of cohortData object

cohortDataShort

a subset of cohortData

cohortDataShortNoCover

a subset of cohortData

species

a data.table with species traits such as longevity, shade tolerance, etc. Must have column speciesCode, with species names/IDs. The following is a list of default trait columns:

"species" same as "speciesCode" -- species ID name

"speciesCode"

"Area" -- inherited from LANDIS-II default table, the Canadian ecoregion from which traits where derived. Not used during the simulation

"firetolerance" -- modulerelative (to other species) fire tolerance

"growthcurve" and "mortalityshape" -- growth curve shape parameters.

"longevity" -- maximum species age

"postfireregen" -- post-fire regeneration strategy ("serotiny", "resprout" or "none")

"resproutprob" -- probability of resprouting

"resproutage_min" -- minimum age at which species is capable of resprouting

"resproutage_max" -- maximum age at which species is capable of resprouting

"seeddistance_eff" -- effective dispersal distance

"seeddistance_max" -- maximum dispersal distance

"shadetolerance" -- relative (to other species) shade tolerance

"sexualmature" -- age at sexual maturity Known optional parameters added/needed by some modules (the user may add others for their own modules):

"inflationFactor" -- Biomass_speciesParameters module: inflation factor for maxB

"growthCurveSource" -- Biomass_speciesParameters module: how "growthcurve" was estimated

"mANPPproportion" -- Biomass_speciesParameters module: multiplication factor to calculate maxANPP from maxB

"thermokarsttol" -- Biomass_disturbances module: proportion of biomass surviving after thermokarst (i.e. permafrost thaw). Applied equally across cohorts. Parameters inherited from LANDIS-II default table, but not used in LandR at the moment:

"leaflongevity"

"wooddecayrate"

"leafLignin"

"hardsoft" Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015) for further detail.

modelCover

statistical model of species presence/absence

modelBiomass

statistical model of species biomass

successionTimestep

The time between successive seed dispersal events. In LANDIS-II, this is called "Succession Timestep".

currentYear

time(sim)

Details

See Details.

establishprob

This section takes the cover as estimated from the mature tree cover and partitions it between resprouting and seeds Unfortunately, establishment by seed is not independent of resprouting, i.e., some pixels would have both Since we don't know the level of independence, we can't correctly assess how much to discount the two. If there is resprouting > 0, then this is the partitioning: establishprob = f(establishprob + resproutprob + jointEstablishProbResproutProb) If jointEstablishProbResproutProb is 0, then these are independent events and the total cover probability can be partitioned easily between seeds and resprout. This is unlikely ever to be the case. We are picking 50% overlap as a number that is better than 0 (totally independent probabilities, meaning no pixel has both seeds and resprout potential) and 100% overlap (totally dependent probabilities, i.e., every pixel where there is seeds will also be a pixel with resprouting) This is expressed with the "* 0.5" in the code.

#' @return A speciesEcoregion data.table with added columns for parameters maxB, maxANPP and establishprob