Simple prepInputs for LCC2005 or LCC2010 data

A wrapper around prepInputs for the Canadian Land Cover Classification product(s).

prepInputsLCC(
  year = 2005,
  destinationPath = asPath("."),
  studyArea = NULL,
  rasterToMatch = NULL,
  filename2 = NULL,
  ...
)

Arguments

year
Numeric, either 2005 or 2010

destinationPath
Character string of a directory in which to download and save the file that comes from url and is also where the function will look for archive or targetFile. NOTE (still experimental): To prevent repeated downloads in different locations, the user can also set options("reproducible.inputPaths") to one or more local file paths to search for the file before attempting to download. Default for that option is NULL meaning do not search locally.

studyArea
SpatialPolygons* object used for masking and possibly cropping if no rasterToMatch is provided. If not in same CRS, then it will be spTransformed to CRS of x before masking. Currently, this function will not reproject the x. Optional in postProcess.

rasterToMatch
Template Raster* object used for cropping (so extent should be the extent of desired outcome) and reprojecting (including changing the resolution and projection). See details in postProcess.

filename2
filename2 is optional, and is either NULL (no writing of outputs to disk), or several options for writing the object to disk. If TRUE (the default), it will give it a file name determined by .prefix(basename(filename1), prefix). If a character string, it will use this as its file name. See determineFilename.

...
Passed to raster::crop