Remove missing cohorts from cohortData based on pixelGroupMap

Remove missing cohorts from cohortData based on pixelGroupMap

rmMissingCohorts(
  cohortData,
  pixelGroupMap,
  cohortDefinitionCols = c("pixelGroup", "age", "speciesCode"),
  doAssertion = getOption("LandR.assertions", TRUE)
)

Arguments

cohortData
A data.table with columns: pixelGroup, ecoregionGroup, speciesCode, and optionally age, B, mortality, aNPPAct, and sumB.

pixelGroupMap
A RasterLayer with pixel values equal to a pixel group number that corresponds exactly to pixelGroup column in cohortData.

cohortDefinitionCols
the columns in cohortData that define unique cohorts

doAssertion
A logical indicating whether some internal tests should be run to ensure the function is running correctly. Default: getOption("LandR.assertions", TRUE).

Value

A list with 2 data.table objects, cohortData and pixelGroupMap, each updated based on missing pixelGroups in the other.