Download and prepare a species traits table for use with Biomass_core module

prepSpeciesTable

getSpeciesTable(url = NULL, dPath = tempdir(), cacheTags = NULL)

prepSpeciesTable(
  speciesTable,
  speciesLayers = NULL,
  sppEquiv = NULL,
  sppEquivCol = "LandR",
  areas = c("BSW", "BP", "MC")
)

Arguments

url

If NULL (the default), uses one from D. Cyr's LANDIS-II files: https://github.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv).

dPath

The destination path.

cacheTags

User tags to pass to Cache.

speciesTable

A species traits table, with at least the following columns:

speciesCode an character representation of species;

Area a character of geographical area within a species range for which trait values are relevant;

longevity species longevity in years, sexualmature age in years at sexual maturity;

shadetolerance numeric value between 1-5 of relative shade tolerance (with respect to other species);

seeddistance_eff the "effective" seed dispersal distance;

seeddistance_max a numeric with the maximum seed dispersal distance;

mortalityshape and growthcurve: growth curve shape parameters. Other columns (e.g. fire-related traits) can also be included depending on LandR modules in use. Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015) for further detail.

speciesLayers

Deprecated.

sppEquiv

table with species name equivalencies between the kNN and final naming formats. See data("sppEquivalencies_CA", "LandR"). For functions that have mixedType, this only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use data("sppEquivalencies_CA", "LandR").

sppEquivCol

the column name to use from sppEquiv. For functions that have mixedType, only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use "Boreal".

areas

A character vector of areas to use. Can be one or more of c("Acadian", "AM", "NorthShore", "BP", "BSE", "BSW", "LSJ", "MC", "PM", "WestON"). If it is more than one, this function will take the minimum value, within a species. These are short versions of the Canada Ecoprovinces. Currently defaults to c("BSW", "BP", "MC") for historical reasons.

Value

getSpeciesTable() returns a data.table with columns LandisCode, Area, Longevity, Maturity, Shade, Fire, SeedEffDist, SeedMaxDist, VegProb, MinAgeVeg, MaxAgeVeg, PostFireRegen, LeafLongevity, WoodDecayRate, MortalityCurve, GrowthCurve, LeafLignin, HardSoft. See assertSpeciesTableRaw() for expected column data types.

prepSpeciesTable() returns a data.table with columns species, Area, longevity, sexualmature, shadetolerance, firetolerance, seeddistance_eff, seeddistance_max, resproutprob, resproutage_min, resproutage_max, postfireregen, leaflongevity, wooddecayrate, mortalityshape, growthcurve, leafLignin, hardsoft. See assertSpeciesTable() for expected column data types.

Note

This one is tailored to Canadian forests

See also