Change species table of parameters/traits

Changes longevity and shade tolerance values in the species table. Longevity values are changed to follow Burton & Cumming (1995) for the following species: Abies balsamea, Abies lasiocarpa, Betula papyrifera, Larix laricina, Larix occidentalis, Picea engelmannii, Picea glauca, Picea mariana, Pinus banksiana, Pinus contorta, Pinus resinosa, Pinus strobus, Populus balsamifera v. balsamifera, Populus tremuloides, Pseudotsuga menziesii var. glauca, Pseudotsuga menziesii, Thuja plicata, Tsuga heterophylla, Tsuga mertensiana x heterophylla, and only for the Boreal Shield West (BSW), Boreal Plains (BP) and Montane Cordillera (MC) speciesTable$Areas. Note that BSW and BP areas correspond more closely to the region considered in Table 2 of Burton & Cumming (1995), while MC will correspond to both tables.

speciesTableUpdate(species, speciesTable, sppEquiv, sppEquivCol)

Arguments

species

a data.table with species traits such as longevity, shade tolerance, etc. Must have column speciesCode, with species names/IDs. The following is a list of default trait columns:

"species" same as "speciesCode" -- species ID name

"speciesCode"

"Area" -- inherited from LANDIS-II default table, the Canadian ecoregion from which traits where derived. Not used during the simulation

"firetolerance" -- modulerelative (to other species) fire tolerance

"growthcurve" and "mortalityshape" -- growth curve shape parameters.

"longevity" -- maximum species age

"postfireregen" -- post-fire regeneration strategy ("serotiny", "resprout" or "none")

"resproutprob" -- probability of resprouting

"resproutage_min" -- minimum age at which species is capable of resprouting

"resproutage_max" -- maximum age at which species is capable of resprouting

"seeddistance_eff" -- effective dispersal distance

"seeddistance_max" -- maximum dispersal distance

"shadetolerance" -- relative (to other species) shade tolerance

"sexualmature" -- age at sexual maturity Known optional parameters added/needed by some modules (the user may add others for their own modules):

"inflationFactor" -- Biomass_speciesParameters module: inflation factor for maxB

"growthCurveSource" -- Biomass_speciesParameters module: how "growthcurve" was estimated

"mANPPproportion" -- Biomass_speciesParameters module: multiplication factor to calculate maxANPP from maxB

"thermokarsttol" -- Biomass_disturbances module: proportion of biomass surviving after thermokarst (i.e. permafrost thaw). Applied equally across cohorts. Parameters inherited from LANDIS-II default table, but not used in LandR at the moment:

"leaflongevity"

"wooddecayrate"

"leafLignin"

"hardsoft" Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015) for further detail.

speciesTable

A species traits table, with at least the following columns:

speciesCode an character representation of species;

Area a character of geographical area within a species range for which trait values are relevant;

longevity species longevity in years, sexualmature age in years at sexual maturity;

shadetolerance numeric value between 1-5 of relative shade tolerance (with respect to other species);

seeddistance_eff the "effective" seed dispersal distance;

seeddistance_max a numeric with the maximum seed dispersal distance;

mortalityshape and growthcurve: growth curve shape parameters. Other columns (e.g. fire-related traits) can also be included depending on LandR modules in use. Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015) for further detail.

sppEquiv

table with species name equivalencies between the kNN and final naming formats. See data("sppEquivalencies_CA", "LandR"). For functions that have mixedType, this only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use data("sppEquivalencies_CA", "LandR").

sppEquivCol

the column name to use from sppEquiv. For functions that have mixedType, only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use "Boreal".

Value

An updated species data.table

Details

Of the above species, shade tolerance values are changed for Abies spp, Picea spp, and Tsuga spp. to reflect western boreal shade tolerances better.

When different longevity/shade tolerance trait values exist for a given species, the minimum value across Area's (BSW, BP, MC) is kept.

ATTENTION: if none of species in species are from BSW, BP or MC area this function will not change any values.

All other species/Area trait values follow Dominic Cyr and Yan Boulanger's trait values available at: (https://raw.githubusercontent.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv).