Table of species name equivalencies for Canadian trees

A table containing the different species names used across diferent sources (e.g., LANDIS-II test parameter files and trait tables, the LandR standard, kNN species biomass layers, ...). Each column refers to a diferent source or species naming approach. Presently only containing Canadian native tree species, with name equivalencies coming from:

LANDIS-II test parameter tables (column LANDIStest_names; source: https://raw.githubusercontent.com/LANDIS-II-Foundation/Extensions-Succession/master/biomass-succession-archive/trunk/tests/v6.0-2.0/species.txt);

LANDIS-II Canada-wide trait table (column LANDIStraits_names; source: https://raw.githubusercontent.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv);

LandR family of SpaDES modules (column LandR_names;

CFS kNN species biomass layers (column KNN_names; source: http://tree.pfc.forestry.ca/kNN-Species.tar);

Canadian Common Attribute Schema for Forest Resource Inventories (column CASFRI_names; source http://www.borealbirds.ca/files/CAS_Document_Final_Mar_2010_ALL_APPENDICES.pdf).

Remaining columns have been filled with some other useful ways to name species (e.g., for plotting).

sppEquivalencies_CA

Format

A data.frame with 271 rows and 10 variables:

LANDIS_test

species names from LANDIS-II test parameter table

LANDIS_traits

species names from LANDIS-II traits parameter table

LandR

species names from LandR modules

KNN

species names from Knn datasets

CASFRI

species names from CASFRI database

Latin_full

accepted species latin names as in http://theplantlist.org

EN_generic_short

Short version of species' common names in English

EN_generic_full

Full species common names in English

Leading

Simple common english names used for leading species

Notes

additional notes and information

Details

This table is currently used as the default equivalencies table in LandR SpaDES modules, but can also serve as a template to costumise species names equivalencies by the user.