Activate serotiny after a (fire) disturbance

Activate serotiny after a (fire) disturbance

doSerotiny(
  burnedPixelCohortData,
  postFirePixelCohortData,
  postFireRegenSummary = NULL,
  species,
  sufficientLight,
  speciesEcoregion,
  currentTime,
  treedFirePixelTableSinceLastDisp,
  calibrate = FALSE
)

Arguments

burnedPixelCohortData
An expanded cohortData data.table with pixel-level cohort information on burnt pixels and columns: severity - fire severity in that pixel calculated based on fire behaviour properties; firetolerance - species-level fire tolerance; severityToleranceDif - the difference between severity and firetolerance.

postFirePixelCohortData
an empty cohortData-like table with columns: age, B, mortality, aNPPAct, and sumB removed; and pixelIndex added.

postFireRegenSummary
a data.table summarizing for which species serotiny/resprouting were activated and in how many pixels, for each year. Only necessary if calibrate = TRUE.

species
a data.table with species traits such as longevity, shade tolerance, etc.

sufficientLight
a data.table containing probability of establishment, given a site's light conditions (X0-X5) for each level of a species shade tolerance (1-5).

speciesEcoregion
A data.table with speciesEcoregion values.

currentTime
integer. The current simulation time obtained with time(sim)

treedFirePixelTableSinceLastDisp
a vector of pixels that burnt and were forested in the previous time step.

calibrate
logical. Determines whether to output postFirePixelCohortData. Defaults to FALSE.

Value

A list of objects: postFirePixelCohortData, a data.table with the cohorts that undergo serotiny; serotinyPixel, a vector of pixels where serotiny was activated; postFireRegenSummary, the updated postFireRegenSummary, if calibrate = TRUE.