Load kNN species layers from online data repository

TODO: description needed

loadkNNSpeciesLayers(
  dPath,
  rasterToMatch = NULL,
  studyArea = NULL,
  sppEquiv,
  year = 2001,
  knnNamesCol = "KNN",
  sppEquivCol = "Boreal",
  thresh = 10,
  url = NULL,
  ...
)

Arguments

dPath

path to the data directory

rasterToMatch

A RasterLayer objects to use as the template for all subsequent raster operations (i.e., the one used throughout the simulation).

studyArea

A SpatialPolygons* object used as the principle study region, passed to reproducible::prepInputs().

sppEquiv

table with species name equivalencies between the kNN and final naming formats. See data("sppEquivalencies_CA", "LandR"). For functions that have mixedType, this only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use data("sppEquivalencies_CA", "LandR").

year

which year's layers should be retrieved? One of 2001 (default) or 2011.

knnNamesCol

character string indicating the column in sppEquiv containing kNN species names. Default "KNN" for when sppEquivalencies_CA is used.

sppEquivCol

the column name to use from sppEquiv. For functions that have mixedType, only necessary if mixedType == 2. If not provided and mixedType == 2, will attempt to use "Boreal".

thresh

the minimum percent cover a species must have (per pixel) to be considered present in the study area. Defaults to 10.

url

the source url for the data, default is KNN 2011 dataset (https://ftp.maps.canada.ca/pub/nrcan_rncan/Forests_Foret/canada-forests-attributes_attributs-forests-canada/2011-attributes_attributs-2011/)

...

Additional arguments passed to reproducible::Cache() and equivalentName(). Also valid: outputPath, and studyAreaName.

Value

A raster stack of percent cover layers by species.