Simple prepInputs for Canadian LCC data

A wrapper around prepInputs for the Canadian Land Cover Classification product(s).

prepInputsLCC(
  year = 2010,
  destinationPath = asPath("."),
  method = c("ngb", "near"),
  filename2 = NULL,
  ...
)

Arguments

year

Numeric, either 2010 or 2015. See note re: backwards compatibility for 2005.

destinationPath

Character string of a directory in which to download and save the file that comes from url and is also where the function will look for archive or targetFile. NOTE (still experimental): To prevent repeated downloads in different locations, the user can also set options("reproducible.inputPaths") to one or more local file paths to search for the file before attempting to download. Default for that option is NULL meaning do not search locally.

method

passed to terra::intersect or raster::intersect, and reproducible::prepInputs

filename2

passed to reproducible::prepInputs

...

Arguments passed to terra::mask (for maskTo), terra::project (for projectTo) or terra::writeRaster (for writeTo) and not used for cropTo, as well postProcess's rasterToMatch and studyArea arguments (see below). Commonly used arguments might be method, touches, and datatype. If filename is passed, it will be ignored; use writeTo = . If reproducible.gdalwarp = TRUE, then these will be passed to the gdal* functions. See them for details.

Note

As of May 2021, NRCAN no longer provides/hosts the LCC2005 data. A privately hosted version of the data is available to maintain backwards compatibility, but new users/projects should use the 2010 (or newer) data.